requirements.txt
matplotlib==3.1.2
numpy==1.18.1
pynrrd==0.4.2
